Moved MenuManagerFLTK to a separate files
